Multiple updates, mostly to correspondance between Chrom and EXp
Parser, swapped stoi to stoll as we are dealing with large files

+++ geneClass/IO/uHeader.cpp
@@ -61,7 +61,7 @@ std::string uHeader::getParam(header_param name) const
 bool uHeader::_validateChrSize(const std::string& sizeString) const
 {
     try {
-        int value = std::stoi(sizeString);
+        long long int value = std::stoll(sizeString);
         /**< Chr size must be above 0 */
         if (value <= 0)
             return false;
